+++ lib/oneview-sdk/resource/api500/c7000/managed_san.rb
@@ -16,6 +16,16 @@ module OneviewSDK
     module C7000
       # Managed SAN resource implementation for API500 C7000
       class ManagedSAN < OneviewSDK::API300::C7000::ManagedSAN
+
+        # Set public attributes
+        # @param [Hash] attributes Public attributes
+        # @option attributes [String] :name
+        # @option attributes [String] :value
+        # @option attributes [String] :valueType
+        # @option attributes [String] :valueFormat
+        def set_public_attributes(attributes)
+          OneviewSDK::API200::ManagedSAN.instance_method(:set_public_attributes).bind(self).call(attributes)
+        end
       end
     end
   end