+++ spec/integration/resource/api300/c7000/switch/delete_spec.rb
@@ -2,12 +2,6 @@ require 'spec_helper'
 
 klass = OneviewSDK::API300::C7000::Switch
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  include_context 'integration api300 context'
-
-  describe '#remove' do
-    it 'remove resource' do
-      @item = klass.find_by($client_300, {}).first
-      expect { @item.remove }.to_not raise_error
-    end
-  end
+  let(:current_client) { $client_300 }
+  include_examples 'SwitchDeleteExample', 'integration api300 context'
 end