+++ spec/integration/resource/api300/synergy/managed_san/delete_spec.rb
@@ -12,17 +12,11 @@
 require 'spec_helper'
 
 klass = OneviewSDK::API300::Synergy::ManagedSAN
-extra_klass = OneviewSDK::API300::Synergy::FCNetwork
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  include_context 'integration api300 context'
+  let(:current_client) { $client_300_synergy }
+  let(:san_manager_ip) { $secrets_synergy['san_manager_ip'] }
+  let(:fc_network_class) { OneviewSDK::API300::Synergy::FCNetwork }
+  let(:fcoe_network_class) { OneviewSDK::API300::Synergy::FCoENetwork }
 
-  describe 'Remove FC Networks' do
-    it 'Remove' do
-      klass.find_by($client_300_synergy, deviceManagerName: $secrets['san_manager_ip']).each do |san|
-        fc = extra_klass.new($client_300_synergy, name: "FC_#{san['name']}")
-        fc.retrieve!
-        fc.delete
-      end
-    end
-  end
+  include_examples 'ManagedSANDeleteExample', 'integration api300 context'
 end