+++ spec/integration/shared_examples/san_manager/create.rb
@@ -1,0 +1,39 @@
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+
+RSpec.shared_examples 'SANManagerCreateExample' do |context_name|
+  include_context context_name
+
+  describe '#create' do
+    it 'raises MethodUnavailable' do
+      item = described_class.new(current_client, {})
+      expect { item.create }.to raise_error(/The method #create is unavailable for this resource/)
+    end
+  end
+
+  describe '#add' do
+    it 'can add resources' do
+      item = OneviewSDK::SANManager.new(current_client)
+      item['providerDisplayName'] = provider_name
+      item['connectionInfo'] = connection_info
+      expect { item.add }.not_to raise_error
+      expect(item['uri']).to be
+    end
+  end
+
+  describe '#self.get_default_connection_info' do
+    it 'Retrieve connection info for provider' do
+      expect { OneviewSDK::SANManager.get_default_connection_info(current_client, SAN_PROVIDER1_NAME) }.to_not raise_error
+    end
+  end
+
+
+end